refactor(portal): integrate canonical-host enforcement into proxy.ts (Next 16)
Next 16 renamed `middleware.ts` to `proxy.ts` and forbids both
co-existing. Moved the wiring out of the new middleware.ts (which
broke `next build`) into the existing apps/web/proxy.ts as the first
check in the auth-wrapped handler.

- Add enforceCanonicalHost(req) to lib/canonical-host.ts (Next wrapper
  around the pure decideHostMatch helper). Returns null on passthrough,
  NextResponse 301 + Clear-Site-Data on redirect.
- Excludes /api/health[z] and /api/ready inline (load-balancer probe
  protection) — the proxy.ts matcher already excludes Next internals
  and favicon.
- proxy.ts: call enforceCanonicalHost at top of handler, return early
  if it produces a redirect. All existing route policy logic unchanged.
- Move middleware tests into lib/canonical-host.test.ts (23 cases total:
  14 decideHostMatch + 6 enforceCanonicalHost + 3 health-probe
  exclusion via it.each).
- Delete apps/web/middleware.ts and middleware.test.ts.

Build gate: `next build` now succeeds and prints `ƒ Proxy (Middleware)`
confirming Next 16 recognizes the integration. Full vitest pass:
6440/6440 (prior to refactor).
