+++ apps/web/lib/canonical-host.test.ts
@@ -1,5 +1,6 @@
-import { describe, expect, it } from "vitest";
-import { decideHostMatch } from "./canonical-host";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
+import { NextRequest } from "next/server";
+import { decideHostMatch, enforceCanonicalHost } from "./canonical-host";
 
 const path = "/foo";
 const search = "?bar=1";
@@ -162,4 +163,113 @@ describe("decideHostMatch", () => {
       targetUrl: "https://portal.example.com/foo",
     });
   });
+});
+
+// Helper: build a NextRequest with optional host / x-forwarded-host headers.
+function buildRequest(opts: {
+  url: string;
+  host?: string;
+  xForwardedHost?: string;
+}): NextRequest {
+  const headers = new Headers();
+  if (opts.host !== undefined) headers.set("host", opts.host);
+  if (opts.xForwardedHost !== undefined) headers.set("x-forwarded-host", opts.xForwardedHost);
+  return new NextRequest(opts.url, { headers });
+}
+
+describe("enforceCanonicalHost (Next.js wrapper)", () => {
+  let originalPublicUrl: string | undefined;
+  let originalAliases: string | undefined;
+
+  beforeEach(() => {
+    originalPublicUrl = process.env.PUBLIC_URL;
+    originalAliases = process.env.PUBLIC_URL_ALIASES;
+  });
+
+  afterEach(() => {
+    if (originalPublicUrl === undefined) delete process.env.PUBLIC_URL;
+    else process.env.PUBLIC_URL = originalPublicUrl;
+    if (originalAliases === undefined) delete process.env.PUBLIC_URL_ALIASES;
+    else process.env.PUBLIC_URL_ALIASES = originalAliases;
+  });
+
+  it("returns null (passthrough) when PUBLIC_URL is unset", () => {
+    delete process.env.PUBLIC_URL;
+    delete process.env.PUBLIC_URL_ALIASES;
+    const req = buildRequest({
+      url: "http://localhost:3000/foo",
+      host: "localhost:3000",
+    });
+    expect(enforceCanonicalHost(req)).toBeNull();
+  });
+
+  it("returns null when host matches canonical", () => {
+    process.env.PUBLIC_URL = "https://portal.example.com";
+    const req = buildRequest({
+      url: "https://portal.example.com/foo",
+      host: "portal.example.com",
+    });
+    expect(enforceCanonicalHost(req)).toBeNull();
+  });
+
+  it('returns 301 redirect with Clear-Site-Data: "storage" for non-canonical host', () => {
+    process.env.PUBLIC_URL = "https://portal.example.com";
+    delete process.env.PUBLIC_URL_ALIASES;
+    const req = buildRequest({
+      url: "http://192.168.1.10:3000/foo",
+      host: "192.168.1.10:3000",
+    });
+    const res = enforceCanonicalHost(req);
+    expect(res).not.toBeNull();
+    expect(res!.status).toBe(301);
+    expect(res!.headers.get("location")).toBe("https://portal.example.com/foo");
+    expect(res!.headers.get("clear-site-data")).toBe('"storage"');
+  });
+
+  it("returns null when host matches PUBLIC_URL_ALIASES entry (LAN bypass)", () => {
+    process.env.PUBLIC_URL = "https://portal.example.com";
+    process.env.PUBLIC_URL_ALIASES = "192.168.1.10:3000";
+    const req = buildRequest({
+      url: "http://192.168.1.10:3000/foo",
+      host: "192.168.1.10:3000",
+    });
+    expect(enforceCanonicalHost(req)).toBeNull();
+  });
+
+  it("prefers x-forwarded-host over host header (reverse-proxy scenario)", () => {
+    process.env.PUBLIC_URL = "https://portal.example.com";
+    delete process.env.PUBLIC_URL_ALIASES;
+    const req = buildRequest({
+      url: "http://localhost:3000/foo",
+      host: "localhost:3000",
+      xForwardedHost: "portal.example.com",
+    });
+    expect(enforceCanonicalHost(req)).toBeNull();
+  });
+
+  it("preserves path and query string in redirect target", () => {
+    process.env.PUBLIC_URL = "https://portal.example.com";
+    delete process.env.PUBLIC_URL_ALIASES;
+    const req = buildRequest({
+      url: "http://localhost:3000/foo?bar=1&baz=2",
+      host: "localhost:3000",
+    });
+    const res = enforceCanonicalHost(req);
+    expect(res!.headers.get("location")).toBe(
+      "https://portal.example.com/foo?bar=1&baz=2",
+    );
+  });
+
+  it.each(["/api/health", "/api/healthz", "/api/ready"])(
+    "never redirects health-probe path %s (load-balancer probe protection)",
+    (probePath) => {
+      process.env.PUBLIC_URL = "https://portal.example.com";
+      delete process.env.PUBLIC_URL_ALIASES;
+      const req = buildRequest({
+        url: `http://192.168.1.10:3000${probePath}`,
+        host: "192.168.1.10:3000",
+      });
+      expect(enforceCanonicalHost(req)).toBeNull();
+    },
+  );
 });
